Moved META_TABLE to Const
